ARM: Multi PBL: Fix image calculation for certain toolchains
The current way to calculate the compressed image position works
with the OSELAS toolchains, but not with the Debian/Ubuntu toolchains.
For these toolchains &image_end_marker already returns the correct
address, adding global_variable_offset() to it is wrong.

The solution seems to be to put the image_end_marker into a different
object file so that the compiler cannot play any tricks to resolve
the address internally in the object file.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>