kfifo: roundup fifo size to next power of two
Comments in include/kfifo.h state that the FIFO size will be rounded up
to the next power of two, but so far we haven't actually done this,
probably because we didn't have roundup_pow_of_two() back then when
kfifo support was added.

Fix that now and do what the comments state.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>