environment: Do not use environment when overlapping with other partitions
Environment partitions are usually specified with their hardcoded offset
and size, either in the device tree or the board file. These partitions
potentially overlap with other partitions read from the partition table.
Overlapping partitions for sure have bad effects. Be more friendly to our
users and warn them when such a situation occurs and stop using that
partition for storing the environment.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>