imx-bbu-nand-fcb: Workaround ROM checking fingerprint before correcting bitflips
The FCB block is protected with a ECC code which can correct 1bit errors
and detect 2bit errors. This works fine for all fields except the
fingerprint marker (Containing "FCB") and the version field. Here the ROM
chooses to check the correct value of the fields *before* running the ECC
correction. Thus a FCB is not used by the ROM anymore when it has a bitflip
in the fingerprint or version. For us this means we have to rewrite the FCB
in this case, so reject the faulty FCB in read_fcb_hamming_13_8() which
triggers a rewrite during the check operation.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>