fs: ubifs: remove not needed code
This patch aggressively removes stuff that we do not need in a readonly
implementation:

- write buffering support
- lpt/ltab code
- garbage collector
- everything under #ifndef __BAREBOX__

This decreases the binary size by about 5k on ARM, but the main reason
for doing this is the idea that things that are not there don't need to
be synced with upstream ubifs code.
The __BAREBOX__ ifdeffery makes the code very hard to read and is a
maintenance burden by itself, so it is removed here aswell.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>