ubifs: Update to v4.18-rc6
This syncs the UBIFS code with Linux-4.19-rc6.

There are many functions in the Linux UBIFS codebase that we do not need
for a readonly implementation. These are missing here, but all removed
functions are annotated as such so it should be relatively easy to copy
a newer codebase over the current one and to see which functions shall
be removed from the newer version.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>