arm: crypto: fix SHA256 shipped assembler code
Binutils 2.29 changed behavior of the adr instruction to always
add the thumb mode switch offset, which breaks usages where the
user is interested in the address of an internal symbol, instead
of a jump address. Binutils 3.31 fixed this by only adding the
offset when interworking is required.

As there are toolchains out there with broken binutils, it's better
to fix the code to work around the issue by not using the named
label.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Tested-by: Robert Schwebel <r.schwebel@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>