phy: call adjust_link in attach if using fixed-link
Normally, phy_update_status is in charge of reporting a change in link
status when phy is updated. When using fixed-link, speed and duplex are
initialized directly after registering the phy and there is no driver.
Hence when phy_update_status is called, the check for new values fails
and returns directly. Since update_link call is mandatory for some
network devices to work, call it directly when using fixed link in
phy_device_attach to force update of link.

Signed-off-by: Clement Leger <clement.leger@kalray.eu>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>