ARM: clk: rk3188: don't set same clk rate twice
We found setting a clock rate which has already been set, rk3188 (radxa rock
pro) bails out. This is a quick fix only. Underlying situation not (yet)
investigated: why it is even trying to set it to the same rate again. It
remains to state that some but not all rrpro boards exhibit this behaviour, no
other rk3188 boards have been tested.

Signed-off-by: P. Rachet <perachet7@gmail.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>