nand: denali: remove undefined GFP_DMA flag
This was a remnant from porting kernel code to barebox.
While being uncritical so far, this will now cause a compiler error
since kzalloc is not a define but a static inline function.

As the kzalloc() 'mode' argument was ignored before and still will be,
it is safe to remove the parameter.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>