fs/fat: fix FAT32 detection
Limits for the number of clusters were used to determine the FAT fs type.
This fails e.g. for FAT32 fs with low cluster number that can be found
in certain Android images.
Sync the FAT fs type detection to the method used in linux by checking
the sectors/FAT entries at offset 0x16 (must be zero for FAT32) and offset
0x24 (must be non-zero for FAT32).

Signed-off-by: Alexander Kurz <akurz@blala.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>