arm: at91: fix clock to mci1 for at91sam9263
at91_add_device_mci() was missing configuration of PIOA6
when configuring mci1.
With this fix we can read data from SD card with at91sam9263ek,
when built without DT.
Building without a DT is required when we do a bootstrap build.

The other at91samxxx_devices was checked - only the 9263 was
missing the CLK configuration for mci1.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>