imx6ul-phytec-phycore-som: don't use fsl, no-blockmark-swap for the NAND
barebox doesn't honor that property which yields to inconsitencies
between Linux and barebox after for example filling an UBI with a file
system with barebox and then trying to mount that from Linux.

Note this is an interruptive change as it requires to rewrite the flash
to make it working with Linux again.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>